Code is Working fine withh OOP. Implementing Return Types
