--- ProjectReport.rtf
@@ -1,8 +1,0 @@
-John Smith
-Franklin Wong
-Joyce English
-Ramesh Narayan
-James Borg
-Jennifer Wallace
-Ahmad Jabbar
-Alicia Zelaya
